+++ Assets/Tests/InputSystem/CoreTests_Actions_Interactions.cs
@@ -32,48 +32,79 @@ internal partial class CoreTests
         {
             trace.SubscribeToAll();
 
+            runtime.currentTime = 1;
             Press(gamepad.buttonSouth);
 
             var actions = trace.ToArray();
-            Assert.That(actions, Has.Length.EqualTo(3));
+            Assert.That(actions, Has.Length.EqualTo(5));
             Assert.That(actions,
                 Has.Exactly(1).With.Property("action").SameAs(pressOnlyAction).And.With.Property("phase")
-                    .EqualTo(InputActionPhase.Performed));
+                    .EqualTo(InputActionPhase.Started).And.With.Property("duration")
+                    .EqualTo(0));
+            Assert.That(actions,
+                Has.Exactly(1).With.Property("action").SameAs(pressOnlyAction).And.With.Property("phase")
+                    .EqualTo(InputActionPhase.Performed).And.With.Property("duration")
+                    .EqualTo(0));
             Assert.That(actions,
                 Has.Exactly(1).With.Property("action").SameAs(pressAndReleaseAction).And.With.Property("phase")
-                    .EqualTo(InputActionPhase.Performed));
+                    .EqualTo(InputActionPhase.Started).And.With.Property("duration")
+                    .EqualTo(0));
+            Assert.That(actions,
+                Has.Exactly(1).With.Property("action").SameAs(pressAndReleaseAction).And.With.Property("phase")
+                    .EqualTo(InputActionPhase.Performed).And.With.Property("duration")
+                    .EqualTo(0));
             Assert.That(actions,
                 Has.Exactly(1).With.Property("action").SameAs(releaseOnlyAction).And.With.Property("phase")
-                    .EqualTo(InputActionPhase.Started));
+                    .EqualTo(InputActionPhase.Started).And.With.Property("duration")
+                    .EqualTo(0));
 
             trace.Clear();
 
+            runtime.currentTime = 2;
             Release(gamepad.buttonSouth);
 
             actions = trace.ToArray();
-            Assert.That(actions, Has.Length.EqualTo(2));
+            Assert.That(actions, Has.Length.EqualTo(3));
             Assert.That(actions,
                 Has.Exactly(1).With.Property("action").SameAs(releaseOnlyAction).And.With.Property("phase")
-                    .EqualTo(InputActionPhase.Performed));
+                    .EqualTo(InputActionPhase.Performed).And.With.Property("duration")
+                    .EqualTo(1));
             Assert.That(actions,
                 Has.Exactly(1).With.Property("action").SameAs(pressAndReleaseAction).And.With.Property("phase")
-                    .EqualTo(InputActionPhase.Performed));
+                    .EqualTo(InputActionPhase.Started).And.With.Property("duration")
+                    .EqualTo(0));
+            Assert.That(actions,
+                Has.Exactly(1).With.Property("action").SameAs(pressAndReleaseAction).And.With.Property("phase")
+                    .EqualTo(InputActionPhase.Performed).And.With.Property("duration")
+                    .EqualTo(0));
 
             trace.Clear();
 
+            runtime.currentTime = 5;
             Press(gamepad.buttonSouth);
 
             actions = trace.ToArray();
-            Assert.That(actions, Has.Length.EqualTo(3));
+            Assert.That(actions, Has.Length.EqualTo(5));
             Assert.That(actions,
                 Has.Exactly(1).With.Property("action").SameAs(pressOnlyAction).And.With.Property("phase")
-                    .EqualTo(InputActionPhase.Performed));
+                    .EqualTo(InputActionPhase.Started).And.With.Property("duration")
+                    .EqualTo(0));
+            Assert.That(actions,
+                Has.Exactly(1).With.Property("action").SameAs(pressOnlyAction).And.With.Property("phase")
+                    .EqualTo(InputActionPhase.Performed).And.With.Property("duration")
+                    .EqualTo(0));
             Assert.That(actions,
                 Has.Exactly(1).With.Property("action").SameAs(pressAndReleaseAction).And.With.Property("phase")
-                    .EqualTo(InputActionPhase.Performed));
+                    .EqualTo(InputActionPhase.Started).And.With.Property("duration")
+                    .EqualTo(0));
+            Assert.That(actions,
+                Has.Exactly(1).With.Property("action").SameAs(pressAndReleaseAction).And.With.Property("phase")
+                    .EqualTo(InputActionPhase.Performed).And.With.Property("duration")
+                    .EqualTo(0));
             Assert.That(actions,
                 Has.Exactly(1).With.Property("action").SameAs(releaseOnlyAction).And.With.Property("phase")
-                    .EqualTo(InputActionPhase.Started));
+                    .EqualTo(InputActionPhase.Started).And.With.Property("duration")
+                    .EqualTo(0));
         }
     }
 
@@ -109,6 +140,7 @@ internal partial class CoreTests
             performedControl = ctx.control;
 
             Assert.That(action.phase, Is.EqualTo(InputActionPhase.Performed));
+            Assert.That(ctx.duration, Is.GreaterThanOrEqualTo(0.4));
         };
         action.started +=
             ctx =>
@@ -118,6 +150,7 @@ internal partial class CoreTests
             startedControl = ctx.control;
 
             Assert.That(action.phase, Is.EqualTo(InputActionPhase.Started));
+            Assert.That(ctx.duration, Is.EqualTo(0.0));
         };
         action.canceled +=
             ctx =>
@@ -127,6 +160,8 @@ internal partial class CoreTests
             canceledControl = ctx.control;
 
             Assert.That(action.phase, Is.EqualTo(InputActionPhase.Canceled));
+            Assert.That(ctx.duration, Is.GreaterThan(0.0));
+            Assert.That(ctx.duration, Is.LessThan(0.4));
         };
         action.Enable();
 
@@ -398,7 +433,7 @@ internal partial class CoreTests
 
             Set(gamepad.leftTrigger, 0.3f);
 
-            Assert.That(trace, Performed<PressInteraction>(pressAction));
+            Assert.That(trace, Started<PressInteraction>(pressAction).AndThen(Performed<PressInteraction>(pressAction)));
 
             trace.Clear();